requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==22.2.0
bcc==0.18.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.0
cairocffi==1.4.0
CairoSVG==2.6.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cssselect2==0.7.0
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.16.1
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
h11==0.14.0
hidpidaemon==18.4.6
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
more-itertools==8.10.0
multidict==6.0.4
netaddr==0.8.0
netifaces==0.11.0
numpy==1.24.2
oauthlib==3.2.0
openai==0.26.5
Pillow==9.4.0
pop-transition==1.1.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.5
pydbus==0.6.0
pygame==2.1.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-debian===0.1.43ubuntu1
python-dotenv==0.21.1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
repolib==2.1.1
repoman==1.4.0
requests==2.25.1
rsa==4.9
SecretStorage==3.3.1
sessioninstaller==0.0.0
six==1.16.0
systemd-python==234
tinycss2==1.2.1
tqdm==4.64.1
typing_extensions==4.5.0
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
uvicorn==0.20.0
wadllib==1.3.6
webencodings==0.5.1
wsproto==1.0.0
xdg==5
xkit==0.0.0
yarl==1.8.2
zipp==1.0.0
